+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## vx.x.x
+ - Use yowasp-yoysys fir synthesis
 
 ## v0.1.0
  * Fix bug when detecting number of modules with yosys 0.9